requirements.txt
Flask
Flask_SQLAlchemy
Flask_Migrate
Flask_JWT_Extended
python-dotenv
PyMySQL
Flask_Login
cryptography
werkzeug
flask-cors
reportlab